Actualizacion de datos y modelos

+++ ApiSuvesaPos/Datos/Models/ClientesFrecuente.cs
@@ -7,7 +7,7 @@ namespace Datos.Models
 {
     public partial class ClientesFrecuente
     {
-        public int Identificacion { get; set; }
+        public long Identificacion { get; set; }
         public string Cedula { get; set; }
         public string Nombre { get; set; }
         public string Observaciones { get; set; }
