+++ ApiSuvesaPos/Datos/Models/KardexConsignacion.cs
@@ -27,6 +27,6 @@ namespace Datos.Models
         public double? ExistAntBod { get; set; }
         public long? IdBodegaDestino { get; set; }
         public long? IdBodegaOrigen { get; set; }
-        public long IdSucursal { get; set; }
+        public int IdSucursal { get; set; }
     }
 }
